backend/requirements.txt
ultralytics
python-multipart
fastapi==0.111.1
uvicorn[standard]==0.23.2
pandas==2.1.1
numpy==1.26.0
tensorflow==2.15.0
scikit-learn==1.3.0
joblib==1.3.2
requests==2.31.0
pydantic==2.6.2
pillow
imagehash
psycopg2-binary
opencv-python
schedule

